Man Push Dum Code BE

+++ FExchange/Controllers/ProductController.cs
@@ -16,6 +16,7 @@ using DataAccess.IRepository;
 using DataAccess.Repository;
 using Microsoft.EntityFrameworkCore;
 using DataAccess.Paging;
+using Microsoft.AspNetCore.Authorization;
 
 namespace FExchange.Controllers
 {
@@ -69,41 +70,52 @@ namespace FExchange.Controllers
 
         }
         [HttpDelete("{id}")]
+        [Authorize(AuthenticationSchemes = "Bearer", Roles = "User")]
         public void delete(int id)
         {
             _postRepository.delete(id);
         }
         [HttpPut("{id}")]
+        [Authorize(AuthenticationSchemes = "Bearer", Roles = "User")]
         public async Task update(int id,[FromForm] ProductPostDTO dto, [FromForm]IFormFile[] files)
         {
             ProductPost product = _postRepository.get(id);
-            product.Description = dto.Description;
-            product.BoughDate = dto.BoughDate;
-            product.Name = dto.Name;
-            product.Price = dto.Price;
-            product.CategoryId = dto.CategoryId;
-            IEnumerable<ProductImage> productImages = _productImageRepository.GetAll(x => x.ProductId == id).ToList();
-            foreach (ProductImage image in productImages)
+            if (dto.Description != null) product.Description = dto.Description;
+            if (dto.BoughtDate != null) product.BoughtDate = dto.BoughtDate;
+            if (dto.Name !=null) product.Name = dto.Name;
+            if (dto.Price != null) product.Price = dto.Price;
+            if (dto.CategoryId != null) product.CategoryId =(int) dto.CategoryId;
+            if (dto.Status != null) product.Status = dto.Status;
+            if (dto.GoodsStatus != null) product.GoodsStatus = dto.GoodsStatus;
+            if (files != null && files.Count()>0)
             {
-                await deleteFile(image.Image);
-                _productImageRepository.delete(image.Id);
-            }
-            foreach(IFormFile file in files)
-            {
-                await uploadFile(file);
-                ProductImage productImage = new ProductImage()
+                IEnumerable<ProductImage> productImages = _productImageRepository.GetAll(x => x.ProductId == id).ToList();
+                foreach (ProductImage image in productImages)
                 {
-                    Image = "https:/merry.blob.core.windows.net/yume/" + file.FileName,
-                    ProductId = id
-                };
-                _productImageRepository.create(productImage);
+                    await deleteFile(image.Image);
+                    _productImageRepository.delete(image.Id);
+                }
+                foreach (IFormFile file in files)
+                {
+                    await uploadFile(file);
+                    ProductImage productImage = new ProductImage()
+                    {
+                        Image = "https:/merry.blob.core.windows.net/yume/" + file.FileName,
+                        ProductId = id
+                    };
+                    _productImageRepository.create(productImage);
+                }
             }
             _postRepository.update(product);
         }
         [HttpPost]
+        [Authorize(AuthenticationSchemes = "Bearer", Roles = "User")]
         public async Task create([FromForm]ProductPostDTO dto, [FromForm] IFormFile[] files)
         {
             ProductPost productPost = _mapper.Map<ProductPost>(dto);
+            //int id = productPost.Id;
+            productPost.Id = 0;
+            productPost.GoodsStatus = 1;
             _postRepository.create(productPost);
             int productID = _postRepository.getMax();
             foreach (IFormFile file in files)
